fix: 페르소나 설정 페이지 몇가지 스타일링 오류 수정, 파일 정돈
